# [WIP] force inlining along method chains of simple arithmetic ops