@dpsanders 2017-06-29
Thanks! Those timings for `*` seem very high. Do you have FMA?