@Sacha0 2017-06-29
> Those timings for * seem very high. Do you have FMA?

Negative: The i7-3520M is Ivy Bridge, and Intel began supporting FMA only with Haswell (Ivy Bridge's successor).

I imagine the speedup should be greater for processors supporting FMA? Timings on your end? Best!