@dpsanders 2017-08-02
Thanks! and sorry for the delay in getting back to this. 

I am seeing the following timings:

- `master`:
```
julia 0.6> @benchmark *(a, b) setup=(a = randinterval(); b = randinterval())
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     24.470 ns (0.00% GC)
  median time:      28.588 ns (0.00% GC)
  mean time:        31.201 ns (0.00% GC)
  maximum time:     170.772 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     996
```

This PR:
```
julia 0.6> @benchmark *(a, b) setup=(a = randinterval(); b = randinterval())
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     20.652 ns (0.00% GC)
  median time:      26.404 ns (0.00% GC)
  mean time:        29.640 ns (0.00% GC)
  maximum time:     289.785 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     997
```
So there seems to be a slight improvement? 
With the other operations it seems to be in the noise.

```
julia 0.6> versioninfo()
Julia Version 0.6.0
Commit 903644385b (2017-06-19 13:05 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin13.4.0)
  CPU: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
  WORD_SIZE: 64
  BLAS: libopenblas (USE64BITINT DYNAMIC_ARCH NO_AFFINITY Haswell)
  LAPACK: libopenblas64_
  LIBM: libopenlibm
  LLVM: libLLVM-3.9.1 (ORCJIT, haswell)
```